Fix user table email field to be unique, delete Sidebar login logout button end the bottom, update UserDropdown log out button event handler
