+++ frontend/src/components/common/UserDropdown.tsx
@@ -8,13 +8,29 @@ import {
 } from "@nextui-org/react";
 import { useRouter } from "next/navigation";
 import { ReactNode } from "react";
+import { useToast } from "@/components/ui/use-toast";
+import Cookies from "js-cookie";
+import { useUserContext } from "@/contexts/user-context";
 
 export default function UserDropdown({ children }: { children: ReactNode }) {
   const router = useRouter();
-
+  const { toast } = useToast();
+  const { setUserContext } = useUserContext();
   const redirectToUserProfile = () => {
     router.push("/user");
   };
+
+  const handleLoggingOut = () => {
+    localStorage.removeItem('userContext');
+    Cookies.remove('token');
+    setUserContext(null);
+    toast({
+      title: "Log out succesfully",
+      description: "see you later!",
+      variant: "success",
+    });
+    router.push('/login');
+  }
 
   return (
     <Dropdown>
@@ -25,7 +41,7 @@ export default function UserDropdown({ children }: { children: ReactNode }) {
         <DropdownItem key="profile" onClick={redirectToUserProfile}>
           User Profile
         </DropdownItem>
-        <DropdownItem key="logout" className="text-danger" color="danger">
+        <DropdownItem key="logout" className="text-danger" color="danger" onClick={handleLoggingOut}>
           Log Out
         </DropdownItem>
       </DropdownMenu>
